requirements.txt
argcomplete==2.0.0
asgiref==3.5.0
bleach==4.1.0
bootpay==1.0.6
boto3==1.21.24
botocore==1.24.24
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.4
coolsms-python-sdk==2.0.3
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-admin-menu==1.6
django-admin-volt==0.0.5
django-elasticsearch-dsl==7.2.2
django-filter==21.1
django-storages==1.12.3
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
durationpy==0.5
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
firebase-admin==5.2.0
future==0.18.2
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpcio==1.46.1
grpcio-status==1.46.1
hjson==3.0.2
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
libsass==0.21.0
msgpack==1.0.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
placebo==0.9.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.63.0
troposphere==3.2.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.0.3
wsgi-request-logger==0.4.6
